Revert "Renamed Project to MERS (Monster Employee Record Sync)"
