+++ ERIS/App.config
@@ -9,7 +9,7 @@
   </connectionStrings>
   <log4net>
     <appender name="FileAppender" type="log4net.Appender.FileAppender,log4net">
-      <file type="log4net.Util.PatternString" value="MERS-%date{yyyy-MM-dd_HH-mm-ss}.log" />
+      <file type="log4net.Util.PatternString" value="ERIS-%date{yyyy-MM-dd_HH-mm-ss}.log" />
       <appendToFile value="true" />
       <layout type="log4net.Layout.PatternLayout">
         <conversionPattern value="[%level] - [Date: %date] [Logger: %logger] - %message %newline" />
