+++ ERIS/Data/SaveData.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MERS.Data
+namespace ERIS.Data
 {
     internal class SaveData
     {
