+++ ERIS/Models/Birth.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MERS.Models
+namespace ERIS.Models
 {
     public class Birth
     {
