+++ ERIS/Process/SendSummary.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MERS.Process
+namespace ERIS.Process
 {
     class SendSummary
     {
